Move `Error` type
Move the `Error` type out of the `linux` module and into the `launch`
module since the errors are associated with that functionality.
